+++ test/scripts/otlp/span-span-attributes.ts
@@ -1,5 +1,5 @@
 import { createTrace } from 'src/trace';
-import { OtlpTransformer } from 'src/transformers/otlp';
+import { OtlpTransformer } from 'src/exporters/otlp';
 import { ATTRIBUTE_NAME, SPAN_NAME } from 'src/utils/constants';
 
 interface Env {
