+++ test/zipkin-exporter.test.ts
@@ -1,4 +1,4 @@
-import { ZipkinJson } from 'src/transformers/zipkin';
+import { ZipkinJson } from 'src/exporters/zipkin';
 import { ATTRIBUTE_NAME, SPAN_NAME } from 'src/utils/constants';
 import { afterAll, afterEach, beforeAll, describe, expect, test } from 'vitest';
 import { UnstableDevWorker } from 'wrangler';
